requirements.txt
asgiref==3.4.1
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.26.4
azure-mgmt-core==1.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.8
django-cors-headers==3.14.0
django-crispy-forms==1.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
protobuf==4.21.9
psycopg2-binary==2.9.5
pytz==2021.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.5.0
urllib3==1.26.15
xlrd==1.2.0
